Include additions from previous live streams
This includes some additional functionality for wrapping the existing
pipe text command that were implemented during a previous live stream.

The goal here is to extend the existing plugin by allowing for gathering
input from the input panel without modifying the existing plugin, or at
least modifying it as little as possible.

For that reason, this includes a wrapper command that can do the
prompting for the command and providing a command history. This thus
also includes a command palette entry for the command as well as
a key binding for working with the history.