Add working_dir
Add in a new argument to set the current working directory. This is named
`working_dir` to match what the `exec` command uses. 

If this argument is not provided, the default is to use the path of the
current file for any view that has a file name associated with it. In
the case that no argument is provided and the current view has no file
name, the working directory is not explicitly set.