Allow execution in place via !cmd
This extends handling so that if the first character of `shell_cmd` or
the first character of the first element of `cmd` is a ! character, the
result of the execution will be inserted into the buffer instead of
replacing the selection.

This allows for executing commands and having their text inlined into
the buffer without clobbering the selection contents.

The selected text is still passed to the external command, but in the
case where there is no selection (or not all selections are non-empty),
the selections are still used as-is instead of using the whole buffer.

This stops one from being able to send an entire buffer to an external
command and insert the result of said command into the middle of the
buffer. This could be extended further to still use that region as the
text, but keep an extra record of where the selections were regardless.