Lock the buffer while the commands are executing
In order to remove any potential surprises, this implements a lock on
the buffer while external commands are executing. 

The buffer is set to be read-only while the background thread is
executing, and then the buffer is put back into the appropriate state 
once all of the commands have finished executing.

This ensures that the state of the buffer remains consistent while all
commands are being executed.